refactor: move bitcoinAddress to lib
